+++ assets/javascripts/kitten/index.js
@@ -61,6 +61,7 @@ export { SimpleCard } from './components/cards/simple-card'
 export { BackerCard } from './components/cards/backer-card'
 export { BackingCard } from './components/cards/backing-card'
 export { CartRewardCard } from './components/cards/cart-reward-card'
+export { ContributionCard } from './components/cards/contribution-card'
 export { RewardEdition } from './components/cards/reward-edition'
 export { TeamCard } from './components/cards/team-card'
 export { Triptych } from './components/cards/triptych'
@@ -131,6 +132,9 @@ export {
 
 // Heroes
 export { Hero } from './components/heroes/hero'
+
+// Illustrations 
+export { RocketIllustration } from './components/illustrations/rocket-illustration'
 
 // Images
 export { ImageCropper } from './components/images/image-cropper'